refactor: 팀 소개 페이지 컴포넌트 경로 변경
